backend/requirements.txt
fastapi
uvicorn
pytesseract
python-multipart
opencv-python
pillow
